% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_turnaround_hours}
\alias{get_turnaround_hours}
\title{Get turnaround time in hours for a given aircraft}
\usage{
get_turnaround_hours(aircraft_name)
}
\arguments{
\item{aircraft_name}{Aircraft name as a character string}
}
\value{
Turnaround time in hours as a numeric value
}
\description{
Get turnaround time in hours for a given aircraft
}
