% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet.R
\name{print.aircraft}
\alias{print.aircraft}
\title{Print an aircraft object}
\usage{
\method{print}{aircraft}(x, ...)
}
\arguments{
\item{x}{An object of class aircraft}

\item{...}{Additional arguments (not used)}
}
\description{
Print an aircraft object
}
